+++ backups/subwindow_lifecycle_controller_pre_pane_toggle_2026-03-05.py
@@ -349,14 +349,6 @@ class SubwindowLifecycleController:
                     except (TypeError, RuntimeError):
                         pass
                     try:
-                        image_viewer.left_pane_toggle_requested.disconnect(app.main_window._toggle_left_pane)
-                    except (TypeError, RuntimeError):
-                        pass
-                    try:
-                        image_viewer.right_pane_toggle_requested.disconnect(app.main_window._toggle_right_pane)
-                    except (TypeError, RuntimeError):
-                        pass
-                    try:
                         image_viewer.about_this_file_requested.disconnect(app._open_about_this_file)
                     except (TypeError, RuntimeError):
                         pass
@@ -393,8 +385,6 @@ class SubwindowLifecycleController:
                 image_viewer.layout_change_requested.connect(app._on_layout_change_requested)
                 image_viewer.privacy_view_toggled.connect(app._on_privacy_view_toggled)
                 image_viewer.smooth_when_zoomed_toggled.connect(app._on_smooth_when_zoomed_toggled)
-                image_viewer.left_pane_toggle_requested.connect(app.main_window._toggle_left_pane)
-                image_viewer.right_pane_toggle_requested.connect(app.main_window._toggle_right_pane)
                 image_viewer.about_this_file_requested.connect(app._open_about_this_file)
                 hist_slot = lambda i=idx: app.dialog_coordinator.open_histogram(i)
                 image_viewer.histogram_requested.connect(hist_slot)